CORTX:33268: Review comment update on libs from APurwa
