+++ libs/ha/ha_common_libs_k8s.py
@@ -29,6 +29,7 @@ import secrets
 import sys
 import time
 from ast import literal_eval
+from http import HTTPStatus
 from time import perf_counter_ns
 
 import yaml
@@ -44,12 +45,13 @@ from commons.utils import config_utils
 from commons.utils import system_utils
 from commons.utils.system_utils import run_local_cmd
 from config import CMN_CFG
+from config import CSM_REST_CFG
 from config import HA_CFG
 from config.s3 import S3_BLKBOX_CFG
 from config.s3 import S3_CFG
+from libs.csm.rest.csm_rest_core_lib import RestClient
 from libs.csm.rest.csm_rest_system_health import SystemHealth
 from libs.di.di_mgmt_ops import ManagementOPs
-from libs.ha.ha_common_api_libs_k8s import HAK8sApiLibs
 from libs.s3.s3_multipart_test_lib import S3MultipartTestLib
 from libs.s3.s3_restapi_test_lib import S3AccountOperationsRestAPI
 from libs.s3.s3_test_lib import S3TestLib
@@ -84,7 +86,7 @@ class HAK8s:
         self.parallel_ios = None
         self.system_random = secrets.SystemRandom()
         self.dir_path = common_const.K8S_SCRIPTS_PATH
-        self.ha_api = HAK8sApiLibs()
+        self.restapi = RestClient(CSM_REST_CFG)
 
     def polling_host(self,
                      max_timeout: int,
@@ -1784,7 +1786,7 @@ class HAK8s:
                     if not header:
                         user = self.mgnt_ops.create_account_users(nusers=1)
                     else:
-                        user = self.ha_api.create_iam_user_with_header(i_i, header)
+                        user = self.create_iam_user_with_header(i_i, header)
                     if user is None:
                         if event.is_set():
                             exp_fail.append(user)
@@ -1807,7 +1809,7 @@ class HAK8s:
                         if not header:
                             resp = self.delete_s3_acc_buckets_objects({user: del_users_dict[user]})
                         else:
-                            resp = self.ha_api.delete_iam_user_with_header(user, header)
+                            resp = self.delete_iam_user_with_header(user, header)
                         if not resp[0]:
                             user_del_failed.append(user)
                             if event.is_set():
@@ -1927,3 +1929,45 @@ class HAK8s:
             pod_ep_dict[pod] = ip_port
 
         return pod_ep_dict
+
+    def create_iam_user_with_header(self, i_d, header):
+        """
+        Function create IAM user with give header info.
+        :param i_d: Int count number for IAM user name creation
+        :param header: Existing header to use for IAM user creation post request
+        :return: None if IAM user REST req fails or Dict response for IAM user successful creation
+        """
+        user = None
+        payload = {}
+        name = f"ha_iam_{i_d}_{time.perf_counter_ns()}"
+        payload.update({"uid": name})
+        payload.update({"display_name": name})
+        LOGGER.info("Creating IAM user request....")
+        endpoint = CSM_REST_CFG["s3_iam_user_endpoint"]
+        resp = self.restapi.rest_call("post", endpoint=endpoint, json_dict=payload,
+                                      headers=header)
+        LOGGER.info("IAM user request successfully sent...")
+        if resp.status_code == HTTPStatus.CREATED:
+            resp = resp.json()
+            user = dict()
+            user.update({resp["keys"][0]["user"]: {
+                "user_name": resp["keys"][0]["user"],
+                "password": S3_CFG["CliConfig"]["s3_account"]["password"],
+                "accesskey": resp["keys"][0]["access_key"],
+                "secretkey": resp["keys"][0]["secret_key"]}})
+        return user
+
+    def delete_iam_user_with_header(self, user, header):
+        """
+        Function delete IAM user with give header info.
+        :param user: IAM user name to be deleted
+        :param header: Existing header to use for IAM user delete request
+        :return: Tuple
+        """
+        endpoint = CSM_REST_CFG["s3_iam_user_endpoint"] + "/" + user
+        LOGGER.info("Sending Delete IAM user request...")
+        response = self.restapi.rest_call("delete", endpoint=endpoint, headers=header)
+        if response.status_code == HTTPStatus.OK:
+            return True, "Deleted user successfully"
+        LOGGER.debug(response.json())
+        return False, response.json()["message"]
